Change edit page fetching method -> migrate to use token for firebase direct access -> change profile picture to webp
