+++ src/app/dashboard/edit/[slug]/page.tsx.bak
@@ -1,0 +1,43 @@
+import { redirect } from "next/navigation";
+import path from "path";
+
+
+
+import NewPage from "./page"
+import { getAllBlogPath, listFolderPaths } from "@/lib/firebaseInterface";
+import { getProtectedBlog } from "@/lib/getProtectedBlog";
+import { verifyTokenServer } from "@/lib/tokenAuth.server";
+import { cookies } from "next/headers";
+
+export default async function Home({ params : promiseParam }: { params: { slug: string } }) {
+
+  const params = await promiseParam;
+
+  const cookieHeader = (await cookies()).get("session")?.value;
+
+  const ok = await verifyTokenServer(cookieHeader);
+
+  if (!ok) {
+    redirect("/login");
+  }
+  const allPath = await getAllBlogPath();
+
+  let goodPath = false;
+
+  for (const e of allPath){
+    if(path.basename(e) !== params.slug) continue;
+    goodPath = true;
+    break;
+  }
+
+  if (!goodPath) {
+    redirect("/dashboard");
+  }
+
+
+  
+  
+  return (
+    <NewPage slug={params.slug} blogDataUrlPair={await getProtectedBlog(params.slug)} />
+  );
+}
