+++ tsconfig.json
@@ -34,7 +34,7 @@
     "**/*.tsx",
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts"
-  ],
+, "src/app/dashboard/edit/[slug]/page.tsx.bak"  ],
   "exclude": [
     "node_modules"
   ]
